Move MainWindow to package
